+++ src/questions/questions.service.ts
@@ -78,6 +78,11 @@ export class QuestionsService {
     if (result.authorId !== user.id) {
       throw new BadRequestException('작성자만 수정이 가능합니다');
     }
+    if (result.accepteAnswerId) {
+      throw new BadRequestException(
+        '답변이 채택된 게시글은 수정이 불가능합니다.',
+      );
+    }
     await this.questionsRepository.save([
       {
         id: result.id,
@@ -98,7 +103,20 @@ export class QuestionsService {
     if (result.authorId !== user.id) {
       throw new BadRequestException('작성자만 수정이 가능합니다');
     }
+    if (result.accepteAnswerId) {
+      throw new BadRequestException(
+        '답변이 채택된 게시글은 삭제가 불가능합니다.',
+      );
+    }
     await this.questionsRepository.delete({ id: questionId });
+    return true;
+  }
+
+  async updateQuestionWatch(questionId: number) {
+    const question = await this.findQuestionOrError(questionId);
+    await this.questionsRepository.save([
+      { id: questionId, watch: question.watch + 1 },
+    ]);
     return true;
   }
 }
